Extracted transfer model into its own module.
